Use potions/regen, many bugfixes - IsOffCooldown now returns "CanUse" instead of "!CanUse" (checking cooldowns should work correctly now) - GameResponseData.Distance is now a float (should no longer see exceptions about converting "dist" to an int) - made some parameter names more consistent - added async call to use a potion - added async call to use regen_hp and regen_mp - CorrectionData now has the correct form. (this was causing character position to reset to (0,0) - onEval now accounts for when strings are null/empty - onEval now handles pot_timeout - when entities are set, type=xy can still contain new entities, this is now accounted for (this was causing monster/player lists to be very wrong) - ALClientSettings.UseDefaultLoggingConfiguration now also logs messages for * namespace - ALClientSettings.SetLogLevel now correctly only sets the logging level for AL.* namespace - added trace logging and error catching for the connection to the new socket.io package - added CancellationTokenSourceExtensions.cs - all usages of token.Cancel have been replaced with the new extension (cancelling a token calls all registered callbacks synchronously, this is undesired behavior) - simplified DynamicDelay - enrichment of gamedata will now consistently ignore "Ignore"d maps - added GItem.Gives (this has data about items that give hp/mp/xp when consumed) - added POT_TIMEOUT regex to RegexCache

+++ AL.SocketClient/SocketModel/GameResponsesData.cs
@@ -36,7 +36,7 @@ namespace AL.SocketClient.SocketModel
         ///     The distance to the entity you are too far away from.
         /// </summary>
         [JsonProperty("dist")]
-        public int Distance { get; init; }
+        public float Distance { get; init; }
 
         /// <summary>
         ///     The amount of gold sent or received.
